rename existing plugins to conform to current naming scheme
