+++ src/vocabularyDefinition.js
